Fixing Cart Functonality With Localstorage

+++ client/src/utils/handleLocalStorage.jsx
@@ -1,6 +1,7 @@
 
-export const addToLocalStorage = (id) => {
-    let cart = {}
+export const addToLocalStorage = (product) => {
+    // console.log(product)
+    let cart = []
 
     // get cart from Localstorage
     const storedCart = localStorage.getItem('cart');
@@ -8,21 +9,33 @@ export const addToLocalStorage = (id) => {
         cart = JSON.parse(storedCart)
     }
     // Add items
-    const quantity = cart[id]
-    if (quantity) {
-        const newQuantity = quantity + 1
-        cart[id] = newQuantity
+    // console.log(cart)
+    const existingProduct = cart.map(data => data._id === product._id)
+    // console.log(existingProduct)
+    if (!existingProduct.length) {
+        cart.push(product)
+        localStorage.setItem('cart', JSON.stringify(cart))
+    } else {
+        let newCart = cart.find(data => data._id === product._id)
+        // console.log(newCart)
+        if (newCart) {
+            cart[newCart.quantity = product.quantity + newCart.quantity]
+            // cart.push(newCart)
+            // console.log(cart)
+            localStorage.setItem('cart', JSON.stringify(cart))
+        } else {
+            cart.push(product)
+            localStorage.setItem('cart', JSON.stringify(cart))
+
+        }
     }
-    else {
-        cart[id] = 1
-    }
-    localStorage.setItem('cart', JSON.stringify(cart))
-    
+
+
 }
 
 
 export const getLocalStorageCart = () => {
-    let cart = {};
+    let cart = [];
 
     //get the shopping cart from local storage
     const storedCart = localStorage.getItem('cart');
